pyproject.toml
[tool.poetry]
name = "ursip"
version = "0.1.0"
description = ""
authors = ["zephyrean <dev@zephyrean.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
pandas = "^2.0.1"
sqlalchemy = "^2.0.13"
openpyxl = "^3.1.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
multi_line_output = 3
force_grid_wrap = 2
include_trailing_comma = true
